xfs: leaf xattrs
- Implemented leaf based extended attributes on-disk headers

- Both version 4 and version 5 attributes can be read now

- Implemented all leaf headers check to ensure metadata integrity

+++ src/add-ons/kernel/file_systems/xfs/ShortAttribute.h
@@ -8,6 +8,7 @@
 
 #include "Attribute.h"
 #include "Inode.h"
+#include "LeafAttribute.h"
 
 
 // xfs_attr_sf_hdr
@@ -51,7 +52,7 @@ private:
 			const char*			fName;
 			AShortFormHeader*	fHeader;
 			AShortFormEntry*	fEntry;
-			uint32				fLastEntryOffset;
+			uint16				fLastEntryOffset;
 };
 
 #endif
